[AIRFLOW-5193] Move GCP Cloud Build to core (#5798)
This commit moves GCP Cloud Build from contrib to core.
For more information check AIP-21.
