+++ docs/howto/operator/gcp/cloud_build.rst
@@ -42,7 +42,7 @@ Build configuration overview
 
 In order to trigger a build, it is necessary to pass the build configuration.
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_cloud_build.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_cloud_build.py
     :language: python
     :dedent: 0
     :start-after: [START howto_operator_gcp_create_build_from_storage_body]
@@ -50,7 +50,7 @@ In order to trigger a build, it is necessary to pass the build configuration.
 
 The source code for the build can come from `Google Cloud Build Storage <https://cloud.google.com/storage/>`__:
 
-.. literalinclude:: ../../../../tests/contrib/operators/test_gcp_cloud_build_operator.py
+.. literalinclude:: ../../../../tests/gcp/operators/test_cloud_build.py
     :language: python
     :dedent: 12
     :start-after: [START howto_operator_gcp_cloud_build_source_gcs_dict]
@@ -58,7 +58,7 @@ The source code for the build can come from `Google Cloud Build Storage <https:/
 
 It is also possible to specify it using the URL:
 
-.. literalinclude:: ../../../../tests/contrib/operators/test_gcp_cloud_build_operator.py
+.. literalinclude:: ../../../../tests/gcp/operators/test_cloud_build.py
     :language: python
     :dedent: 12
     :start-after: [START howto_operator_gcp_cloud_build_source_gcs_url]
@@ -66,7 +66,7 @@ It is also possible to specify it using the URL:
 
 In addition, a build can refer to source stored in `Google Cloud Source Repositories <https://cloud.google.com/source-repositories/docs/>`__.
 
-.. literalinclude:: ../../../../tests/contrib/operators/test_gcp_cloud_build_operator.py
+.. literalinclude:: ../../../../tests/gcp/operators/test_cloud_build.py
     :language: python
     :dedent: 12
     :start-after: [START howto_operator_gcp_cloud_build_source_repo_dict]
@@ -74,7 +74,7 @@ In addition, a build can refer to source stored in `Google Cloud Source Reposito
 
 It is also possible to specify it using the URL:
 
-.. literalinclude:: ../../../../tests/contrib/operators/test_gcp_cloud_build_operator.py
+.. literalinclude:: ../../../../tests/gcp/operators/test_cloud_build.py
     :language: python
     :dedent: 12
     :start-after: [START howto_operator_gcp_cloud_build_source_repo_url]
@@ -89,20 +89,20 @@ Trigger a build
 ^^^^^^^^^^^^^^^
 
 Trigger a build is performed with the
-:class:`~airflow.contrib.operators.gcp_cloud_build_operator.CloudBuildCreateBuildOperator` operator.
+:class:`~airflow.gcp.operators.cloud_build.CloudBuildCreateBuildOperator` operator.
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_cloud_build.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_cloud_build.py
     :language: python
     :dedent: 4
     :start-after: [START howto_operator_create_build_from_storage]
     :end-before: [END howto_operator_create_build_from_storage]
 
 You can use :ref:`Jinja templating <jinja-templating>` with
-:template-fields:`airflow.contrib.operators.gcp_cloud_build_operator.CloudBuildCreateBuildOperator`
+:template-fields:`airflow.gcp.operators.cloud_build.CloudBuildCreateBuildOperator`
 parameters which allows you to dynamically determine values. The result is saved to :ref:`XCom <concepts:xcom>`, which allows it
 to be used by other operators.
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_cloud_build.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_cloud_build.py
     :language: python
     :dedent: 4
     :start-after: [START howto_operator_create_build_from_storage_result]
