+++ tests/gcp/operators/test_cloud_build_system_helper.py
